Initial Commit
This is a work in progress. I am building the GUI for this project from an example file from root.cern.ch called buttonTest.C.
Currently there is no visualization, I will get to that once more of the GUI framework is complete.
There is some discussion as to whether visualization will be 3D or a 2D cross section. The GUI will likely remain the same in either case.